rename serviceWorkerRegistration
